feat: Add implementation of read_from_file function
This commit adds the implementation of the `read_from_file` function in the `main.rs` file. The function takes a file path as a parameter and reads data from the file using the csv crate. Each record is printed to the console. This function will be used for reading data from a file in future development.